# MergeTree derive do_not_merge_across_partitions_select_final setting