--- a/tests/queries/0_stateless/02286_parallel_final.sh
+++ b/tests/queries/0_stateless/02286_parallel_final.sh
@@ -11,7 +11,7 @@ test_random_values() {
     create table tbl_8parts_${layers}granules_rnd (key1 UInt32, sign Int8) engine = CollapsingMergeTree(sign) order by (key1) partition by (key1 % 8);
     insert into tbl_8parts_${layers}granules_rnd select number, 1 from numbers_mt($((layers * 8 * 8192)));
     optimize table tbl_8parts_${layers}granules_rnd final;
-    explain pipeline select * from tbl_8parts_${layers}granules_rnd final settings max_threads = 16;" 2>&1 |
+    explain pipeline select * from tbl_8parts_${layers}granules_rnd final settings max_threads = 16, do_not_merge_across_partitions_select_final = 0;;" 2>&1 |
        grep -c "CollapsingSortedTransform"
 }
 
@@ -25,7 +25,7 @@ test_sequential_values() {
     create table tbl_8parts_${layers}granules_seq (key1 UInt32, sign Int8) engine = CollapsingMergeTree(sign) order by (key1) partition by (key1 / $((layers * 8192)))::UInt64;
     insert into tbl_8parts_${layers}granules_seq select number, 1 from numbers_mt($((layers * 8 * 8192)));
     optimize table tbl_8parts_${layers}granules_seq final;
-    explain pipeline select * from tbl_8parts_${layers}granules_seq final settings max_threads = 8;" 2>&1 |
+    explain pipeline select * from tbl_8parts_${layers}granules_seq final settings max_threads = 8, do_not_merge_across_partitions_select_final = 0;" 2>&1 |
        grep -c "CollapsingSortedTransform"
 }