@kitaisreal 2023-12-25
Automatically derive `do_not_merge_across_partitions_select_final` setting performance improvements:

Create table:
```
DROP TABLE IF EXISTS test_table_partitioned;
CREATE TABLE test_table_partitioned (id UInt64, value String) ENGINE=ReplacingMergeTree ORDER BY id PARTITION BY id % 10 SETTINGS index_granularity = 8192;
INSERT INTO test_table_partitioned SELECT number, number FROM numbers(500000000);
OPTIMIZE TABLE test_table_partitioned;

SELECT name, part_type, marks FROM system.parts WHERE table = 'test_table_partitioned' AND active = 1;

┌─name────────┬─part_type─┬─marks─┐
│ 0_1_4767_5  │ Compact   │  6105 │
│ 1_2_4768_5  │ Compact   │  6105 │
│ 2_3_4769_5  │ Compact   │  6105 │
│ 3_4_4770_5  │ Compact   │  6105 │
│ 4_5_4761_5  │ Compact   │  6105 │
│ 5_6_4762_5  │ Compact   │  6105 │
│ 6_7_4763_5  │ Compact   │  6105 │
│ 7_8_4764_5  │ Compact   │  6105 │
│ 8_9_4765_5  │ Compact   │  6105 │
│ 9_10_4766_5 │ Compact   │  6105 │
└─────────────┴───────────┴───────┘
```

Before:
```
SELECT * FROM test_table_partitioned FINAL FORMAT Null

0 rows in set. Elapsed: 3.937 sec. Processed 501.23 million rows, 12.92 GB (127.30 million rows/s., 3.28 GB/s.)
Peak memory usage: 670.11 MiB.
```

After:
```
SELECT * FROM test_table_partitioned FINAL FORMAT Null

0 rows in set. Elapsed: 0.789 sec. Processed 500.00 million rows, 12.89 GB (633.86 million rows/s., 16.34 GB/s.)
Peak memory usage: 32.00 MiB.
```